updated finding files to file.py
